+++ R/as_index.R
@@ -76,9 +76,6 @@ as_index.matrix <- function(x, ..., chainable = TRUE, contrib = FALSE) {
   storage.mode(x) <- "numeric"
   levels <- if (is.null(rownames(x))) seq_len(nrow(x)) else rownames(x)
   periods <- if (is.null(colnames(x))) seq_len(ncol(x)) else colnames(x)
-  if (any(x <= 0, na.rm = TRUE)) {
-    warning("some elements of 'x' are less than or equal to 0")
-  }
 
   index <- index_skeleton(levels, periods)
   for (t in seq_along(periods)) {